Guardando datos locales antes de la fusion
