+++ nbproject/private/private.properties
@@ -2,5 +2,5 @@ compile.on.save=true
 <<<<<<< HEAD
 user.properties.file=C:\\Users\\Usuario iTC\\AppData\\Roaming\\NetBeans\\17\\build.properties
 =======
-user.properties.file=C:\\Users\\jsbal\\AppData\\Roaming\\NetBeans\\19\\build.properties
+user.properties.file=C:\\Users\\Usuario iTC\\AppData\\Roaming\\NetBeans\\17\\build.properties
 >>>>>>> 8f443599aff9fe284003f0057f9d01204ec33a81
